float to int widget dimensions

+++ include/Widgets/NodeEditor/NodeConnectorLeft.hpp
@@ -7,7 +7,7 @@ class NodeGraphics;
 class NodeConnectorLeft : public QWidget
 {
 public:
-    NodeConnectorLeft(NodeGraphics* parent, NodeInput* input, float y);
+    NodeConnectorLeft(NodeGraphics* parent, NodeInput* input, int y);
 protected:
     NodeInput* input;
     virtual void paintEvent(QPaintEvent*);
