+++ include/Widgets/NodeEditor/NodeConnectorRight.hpp
@@ -7,7 +7,7 @@ class NodeGraphics;
 class NodeConnectorRight : public QWidget
 {
 public:
-    NodeConnectorRight(NodeGraphics* parent, NodeOutput* output, float y);
+    NodeConnectorRight(NodeGraphics* parent, NodeOutput* output, int y);
 protected:
     NodeOutput* output;
     virtual void paintEvent(QPaintEvent*);
